+++ src/theme/variables/testLight.ts
@@ -422,6 +422,9 @@ export const testLight: Theme = {
   guiPluginLogoMoonpay: guiPluginLogoMoonpay,
 
   // UI 4.0:
+
+  badgeDot: palette.accentRed,
+
   // Shadows
   iconShadow: {
     shadowColor: palette.black,
@@ -434,7 +437,7 @@ export const testLight: Theme = {
     elevation: 0
   },
 
-  notifcationCardShadow: {
+  notificationCardShadow: {
     shadowColor: palette.black,
     shadowOffset: {
       width: 3,