+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 ## Unreleased (develop)
 
+## 4.22.0 (staging)
+
+- added: `NotificationCenterScene`
+- added: Search bar to `EarnScene`
+- added: Buy, Sell, Earn, and Trade buttons to `CoinRankingDetailsScene`
+- added: Price chart to `TransactionListScene`
 - added: Add Unizen DEX
+- changed: `TransactionListScene` split into two scenes: `TransactionListScene` and `WalletDetailsScene`
+- changed: All floating `NotificationCard`s are dismissible
+- fixed: Show correct staked balance for deprecated Velodrome pools
 
 ## 4.21.1 (2025-01-28)